Extract longform behavior into Post concern
Move publish!, draft/published predicates, longform_summary, and draft
validation into Posts::Longform to keep Post under the ClassLength limit.
Tighten owner draft-notice assertion to the exact banner copy.
