+++ app/models/concerns/posts/longform.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+# rbs_inline: enabled
+
+# Longform-specific behavior for Post: publishing, draft/published predicates,
+# summary generation, and draft content validation.
+#
+# The post_type/status enums and longform scopes remain on Post itself,
+# since they are referenced as class methods elsewhere.
+module Posts
+  module Longform
+    extend ActiveSupport::Concern
+
+    LONGFORM_SUMMARY_LENGTH = 280
+
+    included do
+      validate :validate_longform_draft_content
+    end
+
+    #: () -> void
+    def publish!
+      self.published_at ||= Time.current
+      self.status = :published
+      save!
+    end
+
+    #: () -> bool
+    def draft_longform?
+      longform? && draft?
+    end
+
+    #: () -> bool
+    def published_longform?
+      longform? && published?
+    end
+
+    #: () -> String
+    def longform_summary
+      stripped = ActionView::Base.full_sanitizer.sanitize(body.to_s).squish
+      stripped.truncate(LONGFORM_SUMMARY_LENGTH)
+    end
+
+    private
+
+    #: () -> void
+    def validate_longform_draft_content
+      return unless draft_longform?
+      return if title.present? || body.present?
+
+      errors.add(:base, I18n.t("posts.longform.errors.blank_draft"))
+    end
+  end
+end
